creacion de hook de session validada
