+++ src/app/agendar/page.tsx
@@ -1,10 +1,11 @@
 import ComeBack from "@/shared/components/ComeBack/comeBack";
 import AppointmentForm from "./appointmentForm";
 
+
 export default function Agendar () {
     return(
         <div className='bg-gradient-to-b from-whiteSmokeBase to-chaarcoalLight min-h-screen flex flex-col justify-center'>
-			<div className='pt-36 md:pt-44 pb-40 w-[90%] md:w-[95%] flex flex-col self-center items-center'>
+			<div className='pt-16 md:pt-28 pb-40 w-[90%] md:w-[95%] flex flex-col self-center items-center'>
 				<div className='relative bottom-6 md:bottom-8 self-start'>
 					<ComeBack/>					
 				</div>
