+++ src/shared/components/navbar/navbar.tsx
@@ -9,25 +9,13 @@ import RegisterButton from '../Buttons/registerButton'
 import LogoNavbar from './logoNavbar'
 import SignOutButton from '../Buttons/signOutButton'
 import { useAppContext } from '@/shared/context'
+import useSession from '@/shared/hooks/useSession'
 
 
 export default function Navbar() {
-	const {userState} = useAppContext()
 	const [menuOpen, setMenuOpen] = useState(false)
 	const pathname = usePathname()
-	const [tokenFromLocalStorage, setTokenFromLocalStorage] = useState('')
-
-	
-
-	useEffect(() => {
-		const token = localStorage.getItem('token')
-		if (token) {
-			setTokenFromLocalStorage(token)
-		} else {
-			setTokenFromLocalStorage('')
-		}
-	}, [userState.token])
-
+	const { loggedIn } = useSession();
 
 	const toggleMenu = () => {
 		setMenuOpen(!menuOpen)
@@ -36,9 +24,6 @@ export default function Navbar() {
 	useEffect(() => {
 		setMenuOpen(false)
 	}, [pathname])
-	
-	
-
 
 	return (
 		<nav className={` flex w-full top-0 z-50 bg-chaarcoal300 py-3 ${menuOpen ? 'h-screen py-4' : 'lg:h-[100px]'}`}>
@@ -80,11 +65,11 @@ export default function Navbar() {
 				{menuOpen && <hr className="border-t border-white w-full mt-2 lg:hidden" />} 
 
 				<div className={`flex-row items-center lg:hidden mt-4 ${menuOpen ? 'flex' : 'hidden'} flex justify-end pb-4`}> 
-					{tokenFromLocalStorage ? <SignOutButton /> : <LoginButton />}
+					{loggedIn ? <SignOutButton /> : <LoginButton />}
 					<RegisterButton />
 				</div>
 				<div className="hidden lg:flex">
-					{tokenFromLocalStorage ? <SignOutButton /> : <LoginButton />}         
+					{loggedIn ? <SignOutButton /> : <LoginButton />}         
 					<RegisterButton />
 				</div>
 			</div>
